Poprawki UI

+++ angular-frontend/src/app/components/comment/comment.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit, Input } from '@angular/core';
 import { Comment } from '../../models/comment.interface';
 import { CommentService } from '../../services/comment.service';
 import { Router } from '@angular/router';
+import { AuthService } from 'src/app/services/auth.service';
 
 @Component({
   selector: 'app-comment',
@@ -11,14 +12,19 @@ import { Router } from '@angular/router';
 export class CommentComponent implements OnInit {
   @Input() comment: Comment;
   @Input() userId: number;
+  commentOwnerName: string;
   isOwner: boolean;
   editing: boolean = false;
   contentEditValue: string;
 
-  constructor(private commentService: CommentService, private router: Router) { }
+  constructor(private commentService: CommentService, private router: Router, private authService: AuthService) { }
 
   ngOnInit() {
     this.checkIfOnwer();
+    this.authService.getUserById(this.comment.user_id)
+      .subscribe(
+        (response: Response) => this.commentOwnerName = response['User']['name']
+      )
   }
 
   private checkIfOnwer() {
@@ -45,12 +51,9 @@ export class CommentComponent implements OnInit {
   }
 
   onDelete() {
-    this.commentService.deleteComment(this.comment.id)
-      .subscribe(
-        () => {
-          this.router.navigate(['/']);
-        }
-      );
+    if(confirm("Na pewno chcesz usunąć ten komentarz?")) {
+      this.commentService.deleteComment(this.comment.id)
+      .subscribe();
+    }
   }
-
 }
